\name{org.Slycopersicum.kozoORGANISM}
\alias{org.Slycopersicum.kozoORGANISM}
\title{The Organism for org.Slycopersicum.kozo}
\description{
  org.Slycopersicum.kozoORGANISM is an R object that contains a single item: a
  character string that names the organism for which org.Slycopersicum.kozo was built.
}
\details{
  Although the package name is suggestive of the organism for which it
  was built, org.Slycopersicum.kozoORGANISM provides a simple way to programmatically
  extract the organism name.
}
\examples{
  org.Slycopersicum.kozoORGANISM
}
\keyword{datasets}

